# backend-platform/services/auth-service/docs/swagger-mobile.yaml
openapi: 3.0.3
info:
  title: Auth Service - Mobile API
  version: "1.0.0"
  description: Mobile/public API surface (authentication, products, orders, user profile)
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authentication response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /api/v1/auth/signup:
    post:
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        roles:
          type: array
          items:
            type: string
        refreshToken:
          type: string
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        countryCode:
          type: string
...